Added state tracking via CSV output.
A CSV output file is placed in the data directory with the timestamp of
the start time in the file name. This tracks all data going through, and
records modifications to the 'state' and 'last contacted' fields.

I also made the data validation in the Sponsor class more strict and
converted the main() fuction into the Main() class.

Some new data validation code is included in libs/validate.py and you
can run this code by running `python madhacksbot.py --validate /path/to.csv`