Added merge tool.
If you have two spreadsheets, a base, and a subset of the base (possibly
with some new rows), the changes from the subset spreadsheet can be
merged into the base (while also adding new rows).

Note, that for a unique ID, the contact name, company name, and contact
email address are all linked together. Changes to these will be migrated
as a new row, so don't touch them. This is mostly intended for migrating
status, last-contact dates, etc.